Move utility funcs into a separate file
