+++ pkg/ctl/create/cluster.go
@@ -10,7 +10,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
-	"github.com/weaveworks/eksctl/pkg/ami"
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha4"
 	"github.com/weaveworks/eksctl/pkg/authconfigmap"
 	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
@@ -101,89 +100,6 @@ func createClusterCmd(g *cmdutils.Grouping) *cobra.Command {
 	group.AddTo(cmd)
 
 	return cmd
-}
-
-// When passing the --without-nodegroup option, don't create nodegroups
-func skipNodeGroupsIfRequested(cfg *api.ClusterConfig) {
-	if withoutNodeGroup {
-		cfg.NodeGroups = nil
-		logger.Warning("cluster will be created without an initial nodegroup")
-	}
-}
-
-// checkEachNodeGroup iterates over each nodegroup and calls check function
-// (this is need to avoid common goroutine-for-loop pitfall)
-func checkEachNodeGroup(cfg *api.ClusterConfig, check func(i int, ng *api.NodeGroup) error) error {
-	for i := range cfg.NodeGroups {
-		if err := check(i, cfg.NodeGroups[i]); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func newNodeGroupChecker(i int, ng *api.NodeGroup) error {
-	if err := api.ValidateNodeGroup(i, ng); err != nil {
-		return err
-	}
-
-	// apply defaults
-	if ng.InstanceType == "" {
-		ng.InstanceType = api.DefaultNodeType
-	}
-	if ng.AMIFamily == "" {
-		ng.AMIFamily = ami.ImageFamilyAmazonLinux2
-	}
-	if ng.AMI == "" {
-		ng.AMI = ami.ResolverStatic
-	}
-
-	if ng.SecurityGroups == nil {
-		ng.SecurityGroups = &api.NodeGroupSGs{
-			AttachIDs: []string{},
-		}
-	}
-	if ng.SecurityGroups.WithLocal == nil {
-		ng.SecurityGroups.WithLocal = api.NewBoolTrue()
-	}
-	if ng.SecurityGroups.WithShared == nil {
-		ng.SecurityGroups.WithShared = api.NewBoolTrue()
-	}
-
-	if ng.AllowSSH {
-		if ng.SSHPublicKeyPath == "" {
-			ng.SSHPublicKeyPath = defaultSSHPublicKey
-		}
-	}
-
-	if ng.VolumeSize > 0 {
-		if ng.VolumeType == "" {
-			ng.VolumeType = api.DefaultNodeVolumeType
-		}
-	}
-
-	if ng.IAM == nil {
-		ng.IAM = &api.NodeGroupIAM{}
-	}
-	if ng.IAM.WithAddonPolicies.ImageBuilder == nil {
-		ng.IAM.WithAddonPolicies.ImageBuilder = api.NewBoolFalse()
-	}
-	if ng.IAM.WithAddonPolicies.AutoScaler == nil {
-		ng.IAM.WithAddonPolicies.AutoScaler = api.NewBoolFalse()
-	}
-	if ng.IAM.WithAddonPolicies.ExternalDNS == nil {
-		ng.IAM.WithAddonPolicies.ExternalDNS = api.NewBoolFalse()
-	}
-
-	return nil
-}
-
-func checkSubnetsGiven(cfg *api.ClusterConfig) bool {
-	return cfg.VPC.Subnets != nil && len(cfg.VPC.Subnets.Private)+len(cfg.VPC.Subnets.Public) != 0
-}
-
-func checkSubnetsGivenAsFlags() bool {
-	return len(*subnets[api.SubnetTopologyPrivate])+len(*subnets[api.SubnetTopologyPublic]) != 0
 }
 
 func doCreateCluster(p *api.ProviderConfig, cfg *api.ClusterConfig, nameArg string, cmd *cobra.Command) error {
